Remove Prometheus Integration Screens
