+++ portal-ui/package.json
@@ -34,7 +34,6 @@
     "minio": "^7.0.32",
     "react": "^18.1.0",
     "react-chartjs-2": "^2.9.0",
-    "react-component-export-image": "^1.0.6",
     "react-copy-to-clipboard": "^5.0.2",
     "react-dom": "^18.1.0",
     "react-dropzone": "^14.2.3",
